--- code/utils2.0.py
@@ -1,147 +1,0 @@
-import os
-import json
-import cv2
-import numpy as np
-from tqdm import tqdm
-
-
-def load_camera_json(json_path: str) -> dict:
-    """
-    Load camera metadata from a JSON file.
-
-    Args:
-        json_path (str): Path to the JSON file.
-
-    Returns:
-        dict: Metadata dictionary extracted from the 'RecordingReport' field in the JSON file.
-    """
-    with open(json_path, 'r') as file:
-        metadata = json.load(file)
-    return metadata.get('RecordingReport', {})
-
-
-def load_metadata_file(root_dir: str) -> dict:
-    """
-    Load the metadata file from the specified directory.
-
-    Args:
-        root_dir (str): Directory where the metadata file is located.
-
-    Returns:
-        dict: Loaded metadata dictionary if found, otherwise None.
-    """
-    metadata_file = 'metadata.nd.json'
-    file_path = os.path.join(root_dir, metadata_file)
-    
-    if os.path.exists(file_path):
-        print(f"Found metadata file at: {file_path}")
-        with open(file_path, 'r') as file:
-            metadata = json.load(file)
-        return metadata
-    
-    print(f"Metadata file {metadata_file} not found in {root_dir}")
-    return None
-
-
-def get_video_paths(directory: str = '/root/capsule/data', subselect: str = None) -> list:
-    """
-    Retrieve video file paths from the specified directory, optionally filtering by a subdirectory.
-
-    Args:
-        directory (str): The directory to search for video files.
-        subselect (str): Optional subdirectory name to filter the search.
-
-    Returns:
-        list: A list of paths to video files.
-    """
-    video_paths = []
-    trial_videos_added = set()  # Track unique trial directories
-
-    for root, _, files in os.walk(directory):
-        if subselect and subselect not in root:
-            continue  # Skip directories that don't match the subselect
-        
-        trial_video_added = False
-        for file in tqdm(files, desc=f"Searching for videos in {root}"):
-            if file.lower().endswith(('.mp4', '.avi')):
-                full_path = os.path.join(root, file)
-                
-                # Ensure only one trial video per directory
-                if 'trial' in file.lower():
-                    if not trial_video_added and root not in trial_videos_added:
-                        video_paths.append(full_path)
-                        trial_video_added = True
-                        trial_videos_added.add(root)
-                else:
-                    print(f"Found video file: {full_path}")
-                    video_paths.append(full_path)
-
-    return video_paths
-
-
-def process_chunk(start: int, chunk_size: int, frame_shape: tuple, video_path: str) -> np.ndarray:
-    """
-    Process a chunk of video frames, converting each to grayscale and resizing.
-
-    Args:
-        start (int): Starting frame index.
-        chunk_size (int): Number of frames to load in the chunk.
-        frame_shape (tuple): Target frame shape as (height, width).
-        video_path (str): Path to the video file.
-
-    Returns:
-        np.ndarray: A stack of processed grayscale frames in the specified chunk.
-
-    Raises:
-        ValueError: If no frames are returned for the specified chunk.
-    """
-    cap = cv2.VideoCapture(video_path)
-    cap.set(cv2.CAP_PROP_POS_FRAMES, start)  # Set starting frame position
-    chunk = []
-
-    for _ in range(chunk_size):
-        ret, frame = cap.read()
-        if not ret:
-            break  # Exit loop if there are no more frames
-        
-        # Convert to grayscale and resize to target shape
-        gray_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-        gray_frame = cv2.resize(gray_frame, (frame_shape[1], frame_shape[0]))
-        chunk.append(gray_frame)
-
-    cap.release()
-    
-    if not chunk:
-        raise ValueError(f"No frames found in chunk starting at {start}. Check the video length.")
-    
-    return np.stack(chunk)
-
-
-def get_results_folder() -> str:
-    """
-    Get the results folder path.
-
-    Returns:
-        str: Path to the results folder.
-    """
-    return '/root/capsule/results'
-
-
-def get_zarr_paths(self, path_to: str = 'gray_frames') -> str:
-    """
-    Construct the path for Zarr storage based on mouse and camera metadata.
-
-    Args:
-        path_to (str): Type of frames to be saved ('gray_frames' or 'motion_energy_frames').
-
-    Returns:
-        str: Full path to the Zarr storage file.
-    """
-    zarr_folder = f"{self.mouse_id}_{self.camera_label}_{self.data_asset_id}"
-    zarr_path = os.path.join(get_results_folder(), zarr_folder)
-    
-    # Create directory if it doesn't exist
-    os.makedirs(zarr_path, exist_ok=True)
-    
-    filename = 'processed_frames.zarr' if path_to == 'gray_frames' else 'motion_energy_frames.zarr'
-    return os.path.join(zarr_path, filename)
